+++ config/file.js
@@ -1,0 +1,23 @@
+const multer = require("multer");
+
+const fileFilter = (req, file, cb) => {
+  if (
+    file.mimetype === "image/jpeg" ||
+    file.mimetype === "image/png" ||
+    file.mimetype === "image/jpg"
+  ) {
+    cb(null, true);
+  } else {
+    // reject file
+    cb({ message: "Unsupported file format" }, false);
+  }
+};
+
+// Multer Configuration
+const upload = multer({
+  dest: "uploads/",
+  limits: { fileSize: 1024 * 1024 },
+  fileFilter: fileFilter,
+});
+
+module.exports = upload;
