+++ example.env
@@ -1,4 +1,8 @@
-PORT=8787
-MONGO_URI=mongodb://localhost:27017/connectmi
-JWT_SECRET=secret
+PORT = 8787
+# MONGO_URI = mongodb://localhost:27017/connectmi
+MONGO_URI = mongodb+srv://e2scoba:blockchain@cluster0.b8qy1lb.mongodb.net/house_linkup?retryWrites=true&w=majority
+JWT_SECRET = secret
 NODE_ENV=development
+CLOUD_NAME = amazing1917
+CLOUD_API_KEY = 466119532983772
+CLOUD_API_SECRET = rXNtC1phVGn79H2FPp4Nr7gOY9M
