+++ routes/health.routes.js
@@ -1,10 +1,7 @@
-// const health = require("../controllers/health.controllers");
-// const healthRoutes = require("express").Router();
+const health = require("../controllers/users/health.controllers");
+const healthRoutes = require("express").Router();
 
-// healthRoutes.get("/health", health.Get);
-// healthRoutes.post("/health", health.Post);
+healthRoutes.get("/health", health.Get);
+healthRoutes.post("/health", health.Post);
 
-// module.exports = healthRoutes;
-const express = require('express');
-
-const router = express.Router();
+module.exports = healthRoutes;
