+++ server.js
@@ -1,41 +1,47 @@
 // importing express framework
-const express = require('express');
+const express = require("express");
 const app = express();
-const cookieParser = require('cookie-parser');
-const normalizeCase = require('./utils/nomaliseCase');
-const userRoutes = require('./routes/users.routes');
+
 // importing .env parser
-const dotenv = require('dotenv');
+const dotenv = require("dotenv");
 dotenv.config();
 
 // importing monogodb database
-const connectDB = require('./config/db');
+const connectDB = require("./config/db");
 connectDB();
 
 // importing middlewares
-const cors = require('cors');
-const bodyParser = require('body-parser');
+const cors = require("cors");
+const bodyParser = require("body-parser");
+const cookieParser = require("cookie-parser");
+const normalizeCase = require("./utils/nomaliseCase");
 // const { protectUser } = require("./middleware/userMiddleware"); // Auth Middlewares
 
-// Routes
-const healthRoutes = require('./routes/health.routes');
+// importing Routes
+const healthRoutes = require("./routes/health.routes");
+const userRoutes = require("./routes/users.routes");
+const profileRoutes = require("./routes/profile.routes");
 
-// Running routes
-app.use(cors({ credentials: true, origin: 'http://localhost:3000' }));
+// Running Global middlewares
+app.use(cors());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 app.use(express.json());
 app.use(cookieParser());
-// app.use('/api/v1/', healthRoutes);
 
-// routes
-// Apply the middleware globally for login and sign-up routes
-app.use(['/api/user/login', '/api/user/signup'], normalizeCase);
-app.use('/api/user', userRoutes);
-// // protected routes
+// Running Routers
+app.use("/api/v1/", healthRoutes);
+
+// middleware to reduce to lowercase
+app.use(["/api/v1/login", "/api/v1/signup"], normalizeCase);
+app.use("/api/v1/", userRoutes);
+
+// protected routes
+// app.use(protectUser);
+app.use("/api/v1/", profileRoutes);
 
 // Error Middlewares
-const { notFound, errorHandler } = require('./middleware/errorMiddleware');
+const { notFound, errorHandler } = require("./middleware/errorMiddleware");
 
 //Not found URL middleware
 app.use(notFound);
